avahi: New env variable to pass a different config file.
This patch adds a new environment variable to the avahi init script
allowing to use a different avahi-daemon.conf config file. This can
be used for testing to stop and restart avahi-daemon restricted to a
single interface, using the "allow-interfaces" option in the config
file. This option can't be passed as a command line argument, so a
new config file needs to be created and passed to the init script.

BUG=chromium:271847
TEST=Manual test, relaunch avahi listening only on wlan0 but not eth0.

Change-Id: I8e663143972de4af2f32ce87a7d52cdf367985de
Reviewed-on: https://chromium-review.googlesource.com/171815
Reviewed-by: David Zeuthen <zeuthen@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>