avahi: Run avahi with "expect daemon".
When avahi is run with --daemonize option it calls fork(2) twice
to launch the "avahi running" process. This requires the upstart
script to include "expect daemon" option instead of "expect fork"
which should be used for daemons that fork only once.

As a side note, the "avahi running" process will fork one more time
to run a "chroot helper" process that is not tracked by upstart, but
that will be killed when SIGKILL is sent to the running process.

BUG=chromium:304237
TEST=Run "start avahi; stop avahi" and doesn't hang.

Change-Id: Idbe0a5de5e63253dc26dd0f30f8477e05cff5d48
Reviewed-on: https://chromium-review.googlesource.com/171883
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: David Zeuthen <zeuthen@chromium.org>
Tested-by: David Zeuthen <zeuthen@chromium.org>
Reviewed-by: Scott James Remnant <keybuk@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>