Use pre-generated perf reports for input perf data
If the perf tool ever changes to become incompatible with the test data,
or if new test data is incompatible with the perf tool, then the tests
will break.

This patch avoids such a failure by adding the pre-generated report
files.  If the reports don't exist, then GetPerfReport will use the perf
tool like before, as in the case of perf files generated by quipper.

BUG=none
TEST=unit tests pass

Change-Id: I7c6daf18f5f51c18a080a843ff03929b7645be68
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/172194