Check event type metadata in greater detail
There are sub-fields within the event type metadata string of the perf
report.  Newer versions of perf report may print more fields.  If the
event type metadata strings are blindly compared, they may mismatch
between versions.

This patch compares them by sub-fields instead.  If a sub-field exists
in one report but not in the other, it does not trigger a mismatch.

BUG=none
TEST=unit tests pass

Change-Id: Id9af5ff6c416e72935b655b58cf0d06fe3ee6e07
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/173561
Reviewed-by: Ahmad Sharif <asharif@chromium.org>