shill: CustomAccessor: Allow custom Clear methods
Allow creators of a CustomAccessor to provide a clear method
instead of using the default of calling Set() with the default
value.  This allows providing a clear method without providing
a set method (not so useful) as well as allowing a property
owner to perform ancillary actions specific to returning a
property to its default state.

BUG=chromium:307692
TEST=Unit test

Change-Id: I02413782e2cccfd6679c3d2945aa218236c82115
Reviewed-on: https://chromium-review.googlesource.com/173321
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: David James <davidjames@chromium.org>