Move stateful update handling to follow stateful wipe.
Previously, a comment said that the code for handling a stateful
update had to precede the stateful wipe.  That comment is no longer
correct.  Moreover, if the code moves to follow the stateful wipe,
it's more logical (why update when a wipe is pending?), and also it
enables a number of code simplifications.

This change contains additional style cleanup, as well.

BUG=None
TEST=Boot a system, run stateful update, run power wash

Change-Id: I3fa3e71fdce6e8a89491307c2b6b7f61f2d72dd2
Reviewed-on: https://chromium-review.googlesource.com/175154
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>