quipper: simplify address mapping code.
Before this CL, PerfParser can look up multiple process mappers for the same
process. Example, if the DSO is not found in the main process mapper, it looks
at the kernel mapper and possibly in the parent's mapper.

Looking at the parent mapper can result in stale information (the parent mapper
could have unmapped this DSO and replaced it with something else).

Having a separate kernel mapper can also lead to complications because kernel
DSOs and user DSOs then have to be kept in separate addresses in quipper space
manually. In fact, AFAICT, the current behavior is wrong and kernel addresses
alias with user addresses in quipper space (look at change:
8b13021e2f47190a645837c9ca3b0a136aa5003b, which introduced the breakage).

With this CL, each process has a single mapper that contains the current state
of the process. Kernel and non-kernel maps both go into the same mapper, and the
AddressMapper takes care of ensuring that DSOs do not get aliased.

BUG=chromium:311712
TEST=Unit tests.
The spew from the unit tests indicates that more events are now mapped (which
were previously being lost because of incorrect mappings).

Change-Id: Id86eac0a60c377bd2b2fd58ec52a7c0e249f903e
Reviewed-on: https://chromium-review.googlesource.com/174685
Reviewed-by: Simon Que <sque@chromium.org>
Commit-Queue: Ahmad Sharif <asharif@chromium.org>
Tested-by: Ahmad Sharif <asharif@chromium.org>