Implemented finalization of endorsement data.
When endorsement data is finalized it only exists in a form that is
bound to a particular CA.  It can no longer be extracted by software
running locally.  Existing endorsement data will be moved into this
state on the next boot.  New endorsement data will be moved into this
state on the next boot after attestation is prepared.

BUG=chromium:243605
TEST=unit, manual

Change-Id: If5adb771de1479ba60298aa9b52378a41e2f47c4
Reviewed-on: https://chromium-review.googlesource.com/176578
Reviewed-by: Darren Krahn <dkrahn@chromium.org>
Commit-Queue: Darren Krahn <dkrahn@chromium.org>
Tested-by: Darren Krahn <dkrahn@chromium.org>