power: Add tests for InputController.
Add tests for the InputController class, which is soon to
get a bit more responsibility. Move more functionality to
InputController::Delegate so that InputController doesn't
call directly into difficult-to-test, interface-less
classes.

Also add an InputInterface interface for the Input class,
along with an InputStub implementation for testing.

Add a bunch of missing 'virtual' keywords to destructors
too.

BUG=chromium:318772
TEST=added new tests; existing tests still pass

Change-Id: Icc42e382640326098ad1bd562e9eea010502fd02
Reviewed-on: https://chromium-review.googlesource.com/176907
Reviewed-by: Daniel Erat <derat@chromium.org>
Tested-by: Daniel Erat <derat@chromium.org>
Commit-Queue: Daniel Erat <derat@chromium.org>