power: Make power_supply_info not print time estimates.
Make power_supply_info avoid printing battery time-to-full
and time-to-empty estimates. These estimates are inaccurate
since they're based on a single reading of the current and
don't match the estimates displayed in the UI or in powerd's
log; they tend to just confuse people.

BUG=chromium:322121
TEST=manual: ran it

Change-Id: I5d1f4a957169463d0119187f96d954faee99d3ef
Reviewed-on: https://chromium-review.googlesource.com/177640
Reviewed-by: Sameer Nanda <snanda@chromium.org>
Commit-Queue: Daniel Erat <derat@chromium.org>
Tested-by: Daniel Erat <derat@chromium.org>