Add and enforce MinorVersion in Payload Manifest.
The MinorVersion is also set to 0 for Full payloads, and to 1 for
DeltaPayloads (for now). If the field is not set, it defaults to 0.

The default is important, since Full payloads will be generated without this
value set in production because older payload generators will be used for
Full payloads for some time to come.

If an unexpected MinorVersion is received, we will refuse to process it, and
send an kErrorCodeUnsupportedMinorPayloadVersion error result.

Add unittests to delta_performer_unittests to unittest the ValidateManifest
method (never individually tested before).

BUG=chromium:322564
TEST=Unittests

Change-Id: Icbd2ebeb739431905497e79edb4b99629c8d6f7f
Reviewed-on: https://chromium-review.googlesource.com/177823
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Commit-Queue: Don Garrett <dgarrett@chromium.org>
Tested-by: Don Garrett <dgarrett@chromium.org>