Update device-wiping comment to be more accurate
This comment used to refer to 'Powerwash', which is a user-facing
feature with a particular UX. The code in question, though, was not
triggering said feature, it was simply performing a similar kind
of stateful partition wipe after a reboot.

I should avoid conflating the two.

BUG=None
TEST=just a comment change.

Change-Id: I4a763fa4676f3a0549374709a0c908f89a6e28eb
Reviewed-on: https://chromium-review.googlesource.com/178080
Commit-Queue: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>