Preserve the .labmachine flag file after stateful update.
The .labmachine file is used to enable certain test lab-specfic
behaviors.  Unfortunately, the file can't be created if the network
fails to come up, and one of the .labmachine behaviors is meant to
repair a network that doesn't come up.

This change ensures that the file is preserved across stateful
updates that otherwise would clobber it.

BUG=chromium:320833
TEST=touch the .labmachine file, stateful update w/ clobber, reboot

Change-Id: Iaa064312814ff7456333f8292c4fed293d562365
Reviewed-on: https://chromium-review.googlesource.com/178204
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>