shill: WiFi: Perform ARP Lookup in PerformTDLSOperation
If the caller specifies an IPv4 address, attempt a IP Address to
MAC translation using the local ARP table, or failing this,
initiate an ARP request for the peer and return an error to the
caller.  While here, return an error if the passed "peer"
parameter looks neither like a MAC address nor an IP address.

BUG=chromium:326518
TEST=Unit tests + updated version of network_WiFi_TDLSPing (CL:179158)

Change-Id: Ib30de44f7d23ee57f5737d1a470ff31eac11b62c
Reviewed-on: https://chromium-review.googlesource.com/179220
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>