crash_sender: upload oldest crashes first
The current logic is sorting crashes by date (which is good), except it
sorts it such that the newest is first (which is bad).

This means if a new process continues to crash, old crash reports will
be perpetually stalled (it's a stack).  Instead, let's start with the
oldest first (like a fifo).

BUG=chromium:326583
TEST=`cbuildbot x86-generic-full` passes VMTests

Change-Id: Ifadedba80cb8c602c094c91d419e4d9645abda51
Reviewed-on: https://chromium-review.googlesource.com/179210
Reviewed-by: Ben Chan <benchan@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>