shill: Device: Refresh Static IP configuration
This CL provides a means for Devices to update static IPConfig
properties in situations where no DHCP server exists.  Previously
IPConfig properties were updated for static IP only when the
IPConfig was initially created, or when the DHCP server returned
a result.  This change allows IPConfigs to be updated whenever
Chrome updates the static IP configuration paramters.

Chrome signals a static IP configuration update by calling the
"Refresh" method on the IPConfig, which will now trigger a
callback in the Device to update the IPConfig.

Methods are added in the IPConfig and StaticIPParameters to
restore the saved (pre-static-IP) configuration before applying
the new static configuration.

BUG=chromium:318290
TEST=Unittest, new autotest network_DhcpFailureWithStaticIP (CL:179783)

Change-Id: I055f27655db7ed137552e9ae8bfa93ba18b88c3d
Reviewed-on: https://chromium-review.googlesource.com/179785
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>