Remove input events from powerd init script
It used to be that powerd's init pre-start script would let powerd take
ownership of all the input device's wakeup sysfs entries.  There is a
little race condition in doing it like this, so udev rules have been set
up to automatically chown the wakeup entries as the devices come up.
This CL depends on the CL's that set that up and just removes the old
style of doing it from the pre-start script.

BUG=chrome-os-partner:24299
TEST=manual testing on Lucas
CQ-DEPEND=Ic6b2687a5bcd499e0baefc935117bb928aaaa1ac

Change-Id: I7bb1a1336f2de7bc87b9f52ae4065ea96eabe933
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/179882
Reviewed-by: Daniel Erat <derat@chromium.org>