Generalize/centralize copying of partitions.
This CL mainly centralizes and generalizes the logic such that we can
treat all partitions as similarly as possible. This means that if a disk
layout creates a FS (say a cache partition) in their disk layout,
that formatted cache partition will also get included on the final
image.

In order to do this in a sane way, I had to remove and refactor a lot
of the code to get rid of the dozens of globals littered throughout the
code. I could continue refactoring but I think the code is much cleaner
at this point.

BUG=chromium:325495
TEST=image_to_usb --install and image_to_usb + boot + chromeos-install

Change-Id: Ie1f1535b8919a33820f89f82009eb7829738a958
Reviewed-on: https://chromium-review.googlesource.com/180040
Tested-by: Chris Sosa <sosa@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Commit-Queue: Chris Sosa <sosa@chromium.org>