Added support for an alternate Privacy CA.
The alternate PCA configuration is an install attribute which means the
lockbox needs to be finalized before attestation initialization can
proceed.  Currently the lockbox waits for attestation work to be done so
this dependency has been reversed.

This CL also takes the first step in updating the dbus interface to
support alternate PCAs.  There will be follow-up CLs in Chrome,
system_api, and cryptohome to finish the change.

BUG=chromium:243605
TEST=unit, manual

Change-Id: I90e65fcce78576e9723f65f5c75d2249e3adc318
Reviewed-on: https://chromium-review.googlesource.com/182541
Reviewed-by: Darren Krahn <dkrahn@chromium.org>
Commit-Queue: Darren Krahn <dkrahn@chromium.org>
Tested-by: Darren Krahn <dkrahn@chromium.org>