Add inhibit jobs to test image
We used to modify upstart jobs when building factory test images.
However, as we are merging factory test image and test image, we need to
inhibit those upstart jobs dynamically. To do so, we are adding
inhibit-* upstart jobs to stop specified upstart jobs when we are
running factory tests.

The inhibited jobs can still be started manually by starting inhibit-*
job with disable_inhibit=1 first.

BUG=chrome-os-partner:24798
TEST=Boot factory test image and see upstart jobs inhibited.

Change-Id: I3182bd8a5df736175a7c075cebe5401fc07823af
Signed-off-by: Vic (Chun-Ju) Yang <victoryang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/181218