power: Create ShutdownReason enum.
This adds a ShutdownReason enum to replace the string that
was passed into Daemon::ShutDown(). Previously, there was a
risk that someone could inadvertently pass an untrusted
string into the function later (which would be bad, since
the string is passed to the shell).

BUG=none
TEST=manual: shutting down still works

Change-Id: I21d149eeaaad4c7ae05cf5edabd8ceea1f0704f5
Reviewed-on: https://chromium-review.googlesource.com/181658
Commit-Queue: Daniel Erat <derat@chromium.org>
Tested-by: Daniel Erat <derat@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>